fix: Trainer-Fly
Heavily based on the code from the pokered wiki.
A bit of a shotgun approach. Essentially, every time a map loads or a battle ends, we set BIT_RESET_NPC_ENCOUNTER. It is only reset when an npc trainer is engaged correctly.
If the npc trainer encounter script is running, and the flag isn't set, we know that trainer-fly occured and abort the battle.